Update and rename pykrita/QuickColor0/QuickColor0.py to pykrita/QuickColor/QuickColor.py
